+++ examples/distributed_simulation/config.py
@@ -9,7 +9,7 @@ from datetime import datetime, timedelta
 
 import pandas as pd
 from dateutil import rrule as rr
-from mango import AgentAddress, activate
+from mango import activate, addr
 
 from assume import World
 from assume.common.market_objects import MarketConfig, MarketProduct
@@ -19,21 +19,20 @@ log = logging.getLogger(__name__)
 
 
 db_uri = os.getenv("DB_URI", "postgresql://assume:assume@localhost:5432/assume")
-db_uri = ""
-use_mqtt = False
+use_mqtt = os.getenv("MQTT_BROKER", False)
 
 tcp_host = os.getenv("TCP_HOST", "0.0.0.0")
-tcp_port = int(os.getenv("TCP_PORT", "9097"))
+tcp_port = int(os.getenv("TCP_PORT", "8089"))
 if use_mqtt:
     manager_protocol_addr = "manager"
-    agent_adress = "agent"
-    agent_adresses = [AgentAddress("agent", "clock_agent")]
+    agent_address = "agent"
+    agent_addresses = [addr("agent", "clock_agent")]
 else:
     manager_protocol_addr = (tcp_host, tcp_port)
-    agent_adress = (tcp_host, 9098)
-    agent_adresses = [AgentAddress((tcp_host, 9098), "clock_agent")]
+    agent_address = (tcp_host, 9098)
+    agent_addresses = [addr((tcp_host, 9098), "clock_agent")]
 
-market_operator_addr = AgentAddress(manager_protocol_addr, "market_operator")
+market_operator_addr = addr(manager_protocol_addr, "market_operator")
 broker_addr = os.getenv("MQTT_BROKER", ("0.0.0.0", 1883, 600))
 
 start = datetime(2019, 1, 1)
@@ -48,10 +47,10 @@ sim_id = "handmade_simulation"
 marketdesign = [
     MarketConfig(
         market_id="EOM",
-        opening_hours=rr.rrule(rr.HOURLY, interval=1, dtstart=start, until=end),
+        opening_hours=rr.rrule(rr.HOURLY, interval=24, dtstart=start, until=end),
         opening_duration=timedelta(hours=1),
         market_mechanism="pay_as_clear",
-        market_products=[MarketProduct(timedelta(hours=1), 1, timedelta(hours=1))],
+        market_products=[MarketProduct(timedelta(hours=1), 24, timedelta(hours=1))],
         additional_fields=["block_id", "link", "exclusive_id"],
     )
 ]
@@ -65,9 +64,6 @@ async def worker(
     n_proc=1,
     m_agents=1,
 ):
-    if world.distributed_role:
-        world.addresses.extend(agent_adresses)
-
     world.setup(
         start=start,
         end=end,
@@ -86,11 +82,12 @@ async def worker(
     if world.distributed_role:
         log.info("sleeping 2s")
         await asyncio.sleep(2)
-        log.info("starting simulation")
+        log.info("starting simulation from %s to %s", start, end)
         await world.async_run(
             start_ts=datetime2timestamp(world.start),
             end_ts=datetime2timestamp(world.end),
         )
     elif world.distributed_role is False:
         async with activate(world.container):
+            log.info("starting worker %s of %s", i + 1, n_proc)
             await world.clock_agent.stopped