+++ tests/test_market.py
@@ -287,6 +287,7 @@ async def test_market_accepted(market_role: MarketRole):
         return all_orders, [], [{"price": 0, "product_start": products[0][0]}], None
 
     market_role.clear = accept_clear
+    # todo add exception handler
 
     accepted, meta = await market_role.clear_market([(start, end, None)])
     assert accepted == orderbook